Move MachRegister into separate files
